# openapi.yaml
openapi: 3.1.0
info:
  title: Silicon Briefcase API
  version: 1.1.0
  description: Organization-scoped file storage, permissions, previews, search, versioning, recovery, and OBO application
    access.
servers:
- url: https://backend.briefcase.teamofsilicons.com/api/v1
  description: The API is served by the backend host. Permanent entry URLs are minted on the application host instead,
    always organization-scoped as https://briefcase.teamofsilicons.com/org/{org_id}/{path}.
security:
- bearerAuth: []
- bearerAuth: []
  testingEnvironmentKey: []
...
paths:
...
  /entries:
    get:
      tags:
      - Entries
      operationId: listEntries
      summary: List folder contents, or filter everything the caller can reach
      description: 'Returns the hundred most recently changed visible entries per page, newest first. Without `filter`,
        the listing browses one level: the organization roots, or the contents of `parent_id`/`path`. With `filter`
        and no parent, it searches every entry the caller may reach, which is how a `location:` predicate selects
        a subtree.


        The filter language combines freely. Terms are ANDed; `or` introduces an alternative; `not` or a leading
        `-` negates; parentheses group. Supported keys:


        `last:N` / `first:N` take N entries chronologically (1-100, single page); `sort:newest` / `sort:oldest`
        order the result, newest first by default; `between:DD-MM-YYYY=DD-MM-YYYY` bounds the last change with both
        days inclusive; `after:DD-MM-YYYY` and `before:DD-MM-YYYY` bound one side; `from:@{carbon:id}` matches the
        creator; `to:@{silicon:id}` matches an explicit share; `for:@{id}` matches what that member can reach; `contains:''term''`
        matches names and extracted content, with `*` as a wildcard; `has:''term''` matches extracted content only;
        `name:''term''` matches names only; `location:''private/cos:tos''` matches a path prefix; `is:` takes `file`,
        `folder`, a renderer (`image`, `video`, `document`, `spreadsheet`, `presentation`, `audio`, `archive`, `code`,
        `unsupported`), or an extension such as `md`; `permissions:` takes `read`, `write`, `update`, `delete`,
        or `manage_permissions` and matches the caller''s effective access. A bare word is shorthand for `contains:`.


        Filtering only ever returns entries the caller can already see.'
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - name: parent_id
        in: query
        schema:
          type: string
          format: uuid
        description: Omit for the organization root
      - name: path
        in: query
        schema:
          type: string
          maxLength: 2048
        description: Parent folder addressed by path instead of identifier
      - name: filter
        in: query
        schema:
          type: string
          maxLength: 1024
        description: Filter expression, for example "last:5 location:'private' (contains:'apple' or contains:'cat')
          is:md"
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/ContentsLimit'
      responses:
        '200':
          description: Visible entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntryPage'
        default:
          $ref: '#/components/responses/Error'
      x-operation-version: 1.0.0
    post:
      tags:
      - Entries
      operationId: createFolder
      description: Without a parent, creates a user folder at the organization base with the required root_type
        access boundary (and IAM tag for type tag). The new folder is a sibling of the reserved containers, not
        redirected into one. With parent_id or parent_path, creates inside that folder and inherits its boundary.
        Existing folders are not relocated. Sibling name collisions return 409 without exposing hidden entry metadata.
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FolderCreate'
      responses:
        '201':
          description: Folder created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        default:
          $ref: '#/components/responses/Error'
      x-operation-version: 1.0.0
  /entries/{entry_id}:
    parameters:
    - $ref: '#/components/parameters/OrgId'
    - $ref: '#/components/parameters/EntryId'
    get:
      tags:
      - Entries
      operationId: getEntry
      responses:
        '200':
          description: Entry metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        default:
          $ref: '#/components/responses/Error'
      x-operation-version: 1.0.0
    patch:
      tags:
      - Entries
      operationId: updateEntry
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                parent_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        default:
          $ref: '#/components/responses/Error'
      x-operation-version: 1.0.0
    delete:
      tags:
      - Entries
      operationId: moveEntryToBin
      responses:
        '204':
          description: Entry moved to the 45-day bin
        default:
          $ref: '#/components/responses/Error'
      x-operation-version: 1.0.0
...
  /bin:
    get:
      tags:
      - Bin
      operationId: listBin
      summary: List recoverable entries, newest deletion first
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/ContentsLimit'
      responses:
        '200':
          description: Deleted entries visible to the actor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntryPage'
        default:
          $ref: '#/components/responses/Error'
      x-operation-version: 1.0.0
...
  /org/{org_id}/{path}:
    servers:
    - url: https://backend.briefcase.teamofsilicons.com
    get:
      tags:
      - Entries
      operationId: resolvePermanentUrl
      summary: Resolve the clean permanent URL of an entry
      description: The path segment mirrors the folder structure exactly, for example `/org/tos/private/cos:tos/top_secret/this_secret.md`.
        The organization segment must match `X-Org-ID`. Without `disposition` the entry and its effective access
        are returned; with `disposition` the file bytes are streamed. Anything the caller cannot read answers `404`.
      parameters:
      - in: query
        name: test_environment
        description: Public sandbox UUID; unavailable or mismatched selectors never fall back to production.
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/OrgId'
      - name: org_id
        in: path
        required: true
        schema:
          type: string
      - name: path
        in: path
        required: true
        schema:
          type: string
          maxLength: 2048
        description: Organization-relative entry path
      - name: disposition
        in: query
        schema:
          type: string
          enum:
          - inline
          - attachment
        description: Omit to receive entry metadata
      - $ref: '#/components/parameters/Range'
      responses:
        '200':
          description: Entry metadata, or complete file bytes when a disposition is requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
            application/octet-stream:
              schema:
                type: string
                format: binary
        '206':
          $ref: '#/components/responses/PartialFileContent'
        default:
          $ref: '#/components/responses/Error'
      x-operation-version: 1.1.0
...
components:
  securitySchemes:
    uploadCapability:
      type: apiKey
      in: header
      name: X-Briefcase-Upload-Capability
      description: Narrow private-byte capability issued by an authorized upload reservation
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: IAM access token
    oboAccess:
      type: apiKey
      in: header
      name: X-IAM-OBO-Access-Proof
      description: Short-lived IAM OBO Access proof
    appId:
      type: apiKey
      in: header
      name: X-App-ID
      description: Originating application identifier required together with an OBO Access proof
    testingEnvironmentKey:
      type: apiKey
      in: header
      name: X-Briefcase-App-Secret
      description: 'IAM test Application secret: ask_ followed by 43 URL-safe characters. Selects an already provisioned
        and registered Briefcase testing environment.'
  parameters:
    OrgId:
      name: X-Org-ID
      in: header
      required: true
      schema:
        type: string
...
    EntryId:
      name: entry_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 255
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
...
    ContentsLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 100
    Range:
      name: Range
      in: header
      description: Single `bytes=` range. A malformed or multi-range value is ignored and the complete file is returned.
      schema:
        type: string
        pattern: ^bytes=(\d+-\d*|-\d+)$
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    PartialFileContent:
      description: Requested byte range of the file
      headers:
        Content-Range:
          schema:
            type: string
          description: bytes {first}-{last}/{complete size}
        Content-Disposition:
          schema:
            type: string
        Accept-Ranges:
          schema:
            const: bytes
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary
  schemas:
...
    ActorRef:
      type: object
      required:
      - type
      - id
      properties:
        type:
          type: string
          enum:
          - carbon
          - silicon
          - application
        id:
          type: string
    FolderCreate:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        parent_id:
          type:
          - string
          - 'null'
          format: uuid
        root_type:
          type: string
          enum:
          - public
          - private
          - tag
          description: Required only when creating at organization root
        tag:
          type: string
          description: Required for a tag root
        invitees:
          type: array
          items:
            $ref: '#/components/schemas/PermissionGrantCreate'
...
    Entry:
      type: object
      required:
      - id
      - org_id
      - type
      - visibility
      - name
      - path
      - permanent_url
      - effective_access
      description: 'A `traversal` entry is a folder the caller reaches only because something inside it was shared.
        It can be opened and lists what was shared; its owner, timestamps, and effective access are withheld. The
        reserved Public, Private, and tag containers carry no `owner`: they are structure IAM reconciliation maintains
        rather than any member''s property.'
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
        type:
          type: string
          enum:
          - file
          - folder
        visibility:
          type: string
          enum:
          - full
          - traversal
        name:
          type: string
        path:
          type: string
          maxLength: 2048
          description: Organization-relative path without a leading separator
        parent_id:
          type:
          - string
          - 'null'
          format: uuid
        root_type:
          type: string
          enum:
          - public
          - private
          - tag
        tag:
          type:
          - string
          - 'null'
        content_type:
          type:
          - string
          - 'null'
        size:
          type:
          - integer
          - 'null'
          minimum: 0
        render:
          description: Renderer a client should open for a file; `unsupported` means metadata and CRUD only.
          oneOf:
          - type: 'null'
          - type: string
            enum:
            - image
            - video
            - document
            - spreadsheet
            - presentation
            - audio
            - archive
            - code
            - unsupported
        permanent_url:
          type: string
          format: uri
          description: Clean permanent URL that shows the folder structure
        content_url:
          type:
          - string
          - 'null'
          format: uri
          description: Authenticated sandboxed content URL for a file
        download_url:
          type:
          - string
          - 'null'
          format: uri
          description: Authenticated attachment URL for a file
        owner:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ActorRef'
        origin_app_id:
          type:
          - string
          - 'null'
        effective_access:
          type: array
          items:
            type: string
            enum:
            - read
            - write
            - update
            - delete
            - manage_permissions
        created_at:
          type:
          - string
          - 'null'
          format: date-time
        updated_at:
          type:
          - string
          - 'null'
          format: date-time
        deleted_at:
          type:
          - string
          - 'null'
          format: date-time
    EntryPage:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Entry'
        next_cursor:
          type:
          - string
          - 'null'
...
    PermissionGrantCreate:
      type: object
      required:
      - principal
      - access
      properties:
        principal:
          $ref: '#/components/schemas/ActorRef'
        access:
          type: array
          minItems: 1
          items:
            type: string
            enum:
            - read
            - write
            - update
        inherit:
          type: boolean
          default: true
...
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
...